# Optimize zsh startup with minimal mode for tmux popups