--- a/tmux/tmux.conf
+++ b/tmux/tmux.conf
@@ -175,14 +175,14 @@ bind-key -n M-q detach-client
 bind-key -n M-w new-window
 
 
-bind-key s display-popup -E "zsh -ic ts"
-bind-key o display-popup -E "zsh -ic to"
-bind-key a display-popup -E "zsh -ic tc"
-bind-key b display-popup -E "zsh -ic tb"
-bind-key -n M-s display-popup -E "zsh -ic ts"
-bind-key -n M-o display-popup -E "zsh -ic to"
-bind-key -n M-a display-popup -E "zsh -ic tc"
-bind-key -n M-b display-popup -E "zsh -ic tb"
+bind-key s display-popup -E "TMUX_POPUP=1 zsh -ic ts"
+bind-key o display-popup -E "TMUX_POPUP=1 zsh -ic to"
+bind-key a display-popup -E "TMUX_POPUP=1 zsh -ic tc"
+bind-key b display-popup -E "TMUX_POPUP=1 zsh -ic tb"
+bind-key -n M-s display-popup -E "TMUX_POPUP=1 zsh -ic ts"
+bind-key -n M-o display-popup -E "TMUX_POPUP=1 zsh -ic to"
+bind-key -n M-a display-popup -E "TMUX_POPUP=1 zsh -ic tc"
+bind-key -n M-b display-popup -E "TMUX_POPUP=1 zsh -ic tb"
 
 
 # Start the minute-aligned time updater (single instance guarded by PID)