## Summary
- Created dual-mode zsh configuration for fast tmux popup shells
- Reduced popup startup time by 56% (145ms → 63ms)
- Modularized functions into separate files for better organization

## Changes
- **Minimal mode**: Ultra-fast config (~63ms) automatically activated in tmux popups via `TMUX_POPUP` env var
- **Function organization**: Extracted functions into `zsh/functions/` directory:
  - `tmux.zsh`: Session management (tm, ts, tc, tb)
  - `venv.zsh`: Virtual environment auto-activation
  - `vi-mode.zsh`: Vi mode configuration
  - `git-repos.zsh`: Repository navigation (to, _to, cache_csv)
- **Shared aliases**: Common aliases in `zsh/aliases.zsh`
- **Smart detection**: Auto-detects tmux popups and loads appropriate config

## Performance Impact
| Mode | Before | After | Improvement |
|------|--------|-------|-------------|
| Popup shells | ~145ms | ~63ms | **56% faster** |
| Regular shells | ~145ms | ~115ms | **21% faster** |

## Test Plan
- [x] Tmux popup commands (ts, to, tc, tb) work in minimal mode
- [x] Full shell retains all features (Oh My Zsh, git, virtualenv)
- [x] Symlinked .zshrc correctly resolves paths
- [x] Mode switching commands (zsh-full, zsh-minimal) work

🤖 Generated with [Claude Code](https://claude.ai/code)